requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
Pydantic
langchain-ollama
duckduckgo_search

